feat: use OPA to check for TF plan changes (#18)
This checks the TF plan output to determine if there are
any resource or output changes.  This is then used to
control if the plan comment should be added to the PR.

Also includes a fix to prevent the terragrunt exit code from
being hidden by the pipe to tee.