feat: alarms for metric traffic problems (#80)
Adds two alarms related to the iOS metrics to staging incident.
1. Catch 4xx errors which can be generated by a bad metrics payload.
2. Catch traffic fluctuations from the expected daily norm.

Highest 4xx spike in past 6 weeks was 1,470, so went with double for the 
threshold.  The goal of the alarm is to catch a case where all phones are 
causing 4xx errors, either from an app or server configuration problem.