fix: increase Prod in-app ECS task memory
The Prod in-app metric ECS task is failing with an OutOfMemory exception.
This increases the CPU and memory of the task to address the error
in the short term.  A long term solution will be required as metrics
will continue to grow.

The ephemeral storage available to the task has also been increased as we
will hit the 20GB default eventually as well.